+++ fltk-sys/src/input.rs
@@ -259,6 +259,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Widget_set_callback(arg1: *mut Fl_Widget, arg2: Fl_Callback, arg3: *mut libc::c_void);
 }
+extern "C" {
+    pub fn Fl_Widget_set_deleter(
+        arg1: *mut Fl_Widget,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Input {
@@ -501,6 +507,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Input_set_callback(arg1: *mut Fl_Input, arg2: Fl_Callback, arg3: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Input_set_deleter(
+        arg1: *mut Fl_Input,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
 }
 extern "C" {
     pub fn Fl_Input_set_value(arg1: *mut Fl_Input, arg2: *const libc::c_char) -> libc::c_int;
@@ -834,6 +846,12 @@ extern "C" {
         arg1: *mut Fl_Int_Input,
         arg2: Fl_Callback,
         arg3: *mut libc::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Int_Input_set_deleter(
+        arg1: *mut Fl_Int_Input,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
 extern "C" {
@@ -1174,6 +1192,12 @@ extern "C" {
         arg1: *mut Fl_Float_Input,
         arg2: Fl_Callback,
         arg3: *mut libc::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Float_Input_set_deleter(
+        arg1: *mut Fl_Float_Input,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
 extern "C" {
@@ -1526,6 +1550,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Multiline_Input_set_deleter(
+        arg1: *mut Fl_Multiline_Input,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
+}
+extern "C" {
     pub fn Fl_Multiline_Input_set_value(
         arg1: *mut Fl_Multiline_Input,
         arg2: *const libc::c_char,
@@ -1875,6 +1905,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Secret_Input_set_deleter(
+        arg1: *mut Fl_Secret_Input,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
+}
+extern "C" {
     pub fn Fl_Secret_Input_set_value(
         arg1: *mut Fl_Secret_Input,
         arg2: *const libc::c_char,
@@ -2212,6 +2248,12 @@ extern "C" {
         arg1: *mut Fl_File_Input,
         arg2: Fl_Callback,
         arg3: *mut libc::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_File_Input_set_deleter(
+        arg1: *mut Fl_File_Input,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
 extern "C" {