+++ fltk-sys/src/valuator.rs
@@ -259,6 +259,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Widget_set_callback(arg1: *mut Fl_Widget, arg2: Fl_Callback, arg3: *mut libc::c_void);
 }
+extern "C" {
+    pub fn Fl_Widget_set_deleter(
+        arg1: *mut Fl_Widget,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Slider {
@@ -501,6 +507,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Slider_set_callback(arg1: *mut Fl_Slider, arg2: Fl_Callback, arg3: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Slider_set_deleter(
+        arg1: *mut Fl_Slider,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
 }
 extern "C" {
     pub fn Fl_Slider_set_bounds(arg1: *mut Fl_Slider, a: f64, b: f64);
@@ -798,6 +810,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Nice_Slider_set_deleter(
+        arg1: *mut Fl_Nice_Slider,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
+}
+extern "C" {
     pub fn Fl_Nice_Slider_set_bounds(arg1: *mut Fl_Nice_Slider, a: f64, b: f64);
 }
 extern "C" {
@@ -1092,6 +1110,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Counter_set_deleter(
+        arg1: *mut Fl_Counter,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
+}
+extern "C" {
     pub fn Fl_Counter_set_bounds(arg1: *mut Fl_Counter, a: f64, b: f64);
 }
 extern "C" {
@@ -1374,6 +1398,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Dial_set_callback(arg1: *mut Fl_Dial, arg2: Fl_Callback, arg3: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Dial_set_deleter(
+        arg1: *mut Fl_Dial,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
 }
 extern "C" {
     pub fn Fl_Dial_set_bounds(arg1: *mut Fl_Dial, a: f64, b: f64);
@@ -1668,6 +1698,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Line_Dial_set_deleter(
+        arg1: *mut Fl_Line_Dial,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
+}
+extern "C" {
     pub fn Fl_Line_Dial_set_bounds(arg1: *mut Fl_Line_Dial, a: f64, b: f64);
 }
 extern "C" {
@@ -1954,6 +1990,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Roller_set_callback(arg1: *mut Fl_Roller, arg2: Fl_Callback, arg3: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Roller_set_deleter(
+        arg1: *mut Fl_Roller,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
 }
 extern "C" {
     pub fn Fl_Roller_set_bounds(arg1: *mut Fl_Roller, a: f64, b: f64);
@@ -2245,6 +2287,12 @@ extern "C" {
         arg1: *mut Fl_Scrollbar,
         arg2: Fl_Callback,
         arg3: *mut libc::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Scrollbar_set_deleter(
+        arg1: *mut Fl_Scrollbar,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
 extern "C" {
@@ -2540,6 +2588,12 @@ extern "C" {
         arg1: *mut Fl_Value_Slider,
         arg2: Fl_Callback,
         arg3: *mut libc::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Value_Slider_set_deleter(
+        arg1: *mut Fl_Value_Slider,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
 extern "C" {
@@ -2842,6 +2896,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Adjuster_set_deleter(
+        arg1: *mut Fl_Adjuster,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
+}
+extern "C" {
     pub fn Fl_Adjuster_set_bounds(arg1: *mut Fl_Adjuster, a: f64, b: f64);
 }
 extern "C" {
@@ -3134,6 +3194,12 @@ extern "C" {
         arg1: *mut Fl_Value_Input,
         arg2: Fl_Callback,
         arg3: *mut libc::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Value_Input_set_deleter(
+        arg1: *mut Fl_Value_Input,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
 extern "C" {
@@ -3431,6 +3497,12 @@ extern "C" {
         arg1: *mut Fl_Value_Output,
         arg2: Fl_Callback,
         arg3: *mut libc::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Value_Output_set_deleter(
+        arg1: *mut Fl_Value_Output,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
 extern "C" {
@@ -3736,6 +3808,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Fill_Slider_set_deleter(
+        arg1: *mut Fl_Fill_Slider,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
+    );
+}
+extern "C" {
     pub fn Fl_Fill_Slider_set_bounds(arg1: *mut Fl_Fill_Slider, a: f64, b: f64);
 }
 extern "C" {
@@ -4027,6 +4105,12 @@ extern "C" {
         arg1: *mut Fl_Fill_Dial,
         arg2: Fl_Callback,
         arg3: *mut libc::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Fill_Dial_set_deleter(
+        arg1: *mut Fl_Fill_Dial,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
 extern "C" {
@@ -4322,6 +4406,12 @@ extern "C" {
         arg1: *mut Fl_Hor_Slider,
         arg2: Fl_Callback,
         arg3: *mut libc::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Hor_Slider_set_deleter(
+        arg1: *mut Fl_Hor_Slider,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
 extern "C" {
@@ -4626,6 +4716,12 @@ extern "C" {
         arg1: *mut Fl_Hor_Fill_Slider,
         arg2: Fl_Callback,
         arg3: *mut libc::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Hor_Fill_Slider_set_deleter(
+        arg1: *mut Fl_Hor_Fill_Slider,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
 extern "C" {
@@ -4937,6 +5033,12 @@ extern "C" {
         arg1: *mut Fl_Hor_Nice_Slider,
         arg2: Fl_Callback,
         arg3: *mut libc::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Hor_Nice_Slider_set_deleter(
+        arg1: *mut Fl_Hor_Nice_Slider,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
 extern "C" {
@@ -5257,6 +5359,12 @@ extern "C" {
         arg1: *mut Fl_Hor_Value_Slider,
         arg2: Fl_Callback,
         arg3: *mut libc::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Hor_Value_Slider_set_deleter(
+        arg1: *mut Fl_Hor_Value_Slider,
+        arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
 extern "C" {