+++ fltk/src/dialog.rs
@@ -502,9 +502,28 @@ impl FileChooser {
             let callback: Option<
                 unsafe extern "C" fn(arg1: *mut Fl_File_Chooser, data: *mut raw::c_void),
             > = Some(shim);
-            Fl_File_Chooser_callback(self._inner, callback, data)
+            Fl_File_Chooser_set_callback(self._inner, callback, data)
         }
     }
+
+        /// Sets the callback of the FileChooser
+        pub fn set_callback2<F: FnMut(&mut Self) + 'static>(&mut self, cb: F) {
+            assert!(!self._inner.is_null());
+            unsafe {
+                unsafe extern "C" fn shim(arg1: *mut Fl_File_Chooser, data: *mut raw::c_void) {
+                    let mut wid = FileChooser { _inner: arg1 };
+                    let a: *mut Box<dyn FnMut(&mut FileChooser)> = data as *mut Box<dyn FnMut(&mut FileChooser)>;
+                    let f: &mut (dyn FnMut(&mut FileChooser)) = &mut **a;
+                    let _ = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| f(&mut wid)));
+                }
+                let a: *mut Box<dyn FnMut(&mut Self)> = Box::into_raw(Box::new(Box::new(cb)));
+                let data: *mut raw::c_void = a as *mut raw::c_void;
+                let callback: Option<
+                    unsafe extern "C" fn(arg1: *mut Fl_File_Chooser, data: *mut raw::c_void),
+                > = Some(shim);
+                Fl_File_Chooser_set_callback(self._inner, callback, data)
+            }
+        }
 
     /// Sets the color of the FileChooser
     pub fn set_color(&mut self, c: Color) {
@@ -733,13 +752,6 @@ impl FileChooser {
     /// Gets the user data of the FileChooser
     /// # Safety
     /// Can invalidate the user data while the FileChooser is in use
-    pub unsafe fn raw_user_data(&self) -> *mut raw::c_void {
-        Fl_File_Chooser_user_data(self._inner)
-    }
-
-    /// Gets the user data of the FileChooser
-    /// # Safety
-    /// Can invalidate the user data while the FileChooser is in use
     pub unsafe fn user_data(&self) -> Option<Box<dyn FnMut()>> {
         let ptr = Fl_File_Chooser_user_data(self._inner);
         if ptr.is_null() {
@@ -747,7 +759,7 @@ impl FileChooser {
         } else {
             let x = ptr as *mut Box<dyn FnMut()>;
             let x = Box::from_raw(x);
-            Fl_File_Chooser_callback(self._inner, None, std::ptr::null_mut());
+            Fl_File_Chooser_set_callback(self._inner, None, std::ptr::null_mut());
             Some(*x)
         }
     }
@@ -757,6 +769,17 @@ impl FileChooser {
     /// Can invalidate the user data while the FileChooser is in use
     pub unsafe fn set_user_data(&mut self, d: *mut raw::c_void) {
         Fl_File_Chooser_set_user_data(self._inner, d)
+    }
+
+    /// Manually unset a callback
+    /// # Safety
+    /// Invoking the callback after being unset is undefined
+    pub unsafe fn unset_callback(&mut self) {
+        let old_data = self.user_data();
+        if old_data.is_some() {
+            self.set_user_data(std::ptr::null_mut() as *mut raw::c_void);
+            let _old_data = old_data.unwrap();
+        }
     }
 
     /// Gets the file or dir name chosen by the FileChooser